cleanup
put the Exercise and the Set class in their own files, also removed dead code from Page3.java
